+++ src/components/setting/enums.ts
@@ -1,0 +1,7 @@
+// eslint-disable-next-line no-shadow
+enum SETTING_TAB {
+  GENERAL = '1',
+  TOOLBOX = '2',
+}
+
+export default SETTING_TAB;
